refactor: consolidate secp256k1 extensions
